Adapted htmLawed rules to prevent clickjacking
HTML forms can be used to trigger a HTTP POST request to an arbitrary URL, without the user knowing about it beforehand. In a similar fashion, one can use the href attribute to create hyperlinks that are misleading. Furthermore, the whitelisted values for the style attribute can be abused to manipulate the page layout in a way that e.g. the page footer is not rendered by the browser anymore, making several elements inaccessible. This was possible by setting the font-size to extremely large values.